Compliance with cockroachdb / postgresql, use werkzeug filesystem cache

+++ app/model.py
@@ -60,7 +60,7 @@ class MachineInterface(Base):
 
     mac = Column(String(17), nullable=False)
     name = Column(String, nullable=False)
-    netmask = Column(SmallInteger, nullable=False)
+    netmask = Column(Integer, nullable=False)
     ipv4 = Column(String(15), nullable=False)
     cidrv4 = Column(String(15 + 3), nullable=False)
     as_boot = Column(Boolean, default=False)
